\docType{import}
\name{reexports}
\alias{reexports}
\alias{train}
\alias{components}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{train}}}
  \item{generics}{\code{\link[generics]{components}}}
}}
